fix: don't do a full verify before archiving.

+++ readonlyhash.sh
@@ -311,16 +311,25 @@ archive_directory() {
 		return 0
 	fi
 
+	# local tmp=$(find "$dir" -name '*.sha256' -type f -not -path '*/.roh.git/*')
+	# echo "$tmp"
+
+	# verify_directory "$dir"
 	if [ ! -d "$ROH_DIR/.git" ]; then
 		echo "ERROR: local repo [$ROH_DIR/.git] does not exist"
 		echo
 		exit 1
 	fi
 
-	# local tmp=$(find "$dir" -name '*.sha256' -type f -not -path '*/.roh.git/*')
-	# echo "$tmp"
+	git_status=$($GIT_BIN -C "$dir" status)
+	echo "$git_status"
+	if ! [[ "$git_status" =~ "nothing to commit, working tree clean" ]]; then
+		echo
+        echo "ERROR: local repo [$ROH_DIR] not clean"
+		echo
+		exit 1
+	fi
 
-	verify_directory "$dir"
 	if [ -f "$dir/_.roh.sqlite3" ]; then
 		rm -r "$dir/.roh.sqlite3"
 	fi
